Lay groundwork for an abstract base class that allows modules to listen to graphics events without the graphics library having to know about the module class.
Eventually, the graphics objects should be passing themselves to the handle_event function associated with their listener, and that listener should determine what to do based on the type of graphics object and its name, among other things. This will allow the graphics code to be entirely isolated from the rest of the code and compiled as a standalone library. Also fix up some headers and make proper use of virtual functions in this commit.